+++ test/integration/API.test.ts
@@ -9,7 +9,7 @@ import chai, { assert, expect } from 'chai'
 chai.use(require('chai-as-promised'))
 import { randomBytes } from 'crypto'
 
-import { ConvexAccount, ConvexAPI } from '../../src'
+import { API, KeyPair } from '../../src'
 import { isAddress, toAddress } from '../../src/Utils'
 
 const PRIVATE_TEST_KEY_TEXT = `
@@ -26,20 +26,21 @@ const PRIVATE_TEST_KEY_PASSWORD = 'secret'
 const CONVEX_URL = 'https://convex.world'
 const TOPUP_AMOUNT = 10000000
 
-describe('ConvexAPI Class', () => {
+describe('API Class', () => {
     describe('requestFunds', async () => {
         it('should request funds from a new account', async () => {
-            const convex = new ConvexAPI(CONVEX_URL)
-            const account = await convex.createAccount()
+            const convex = API.create(CONVEX_URL)
+            const keyPair = KeyPair.create()
+            const account = await convex.createAccount(keyPair)
             const amount = 1000
             const result = await convex.requestFunds(amount, account)
             assert.equal(result, amount)
         })
 
         it('should request funds from the test account', async () => {
-            const convex = new ConvexAPI(CONVEX_URL)
-            const importAccount = ConvexAccount.importFromString(PRIVATE_TEST_KEY_TEXT, PRIVATE_TEST_KEY_PASSWORD)
-            const account = await convex.createAccount(importAccount)
+            const convex = API.create(CONVEX_URL)
+            const importKeyPair = KeyPair.importFromString(PRIVATE_TEST_KEY_TEXT, PRIVATE_TEST_KEY_PASSWORD)
+            const account = await convex.createAccount(importKeyPair)
             const amount = 1000000
             const result = await convex.requestFunds(amount, account)
             assert.equal(result, amount)
@@ -48,9 +49,9 @@ describe('ConvexAPI Class', () => {
 
     describe('query', async () => {
         it('should throw a ConvexAPIRequestError on a syntax error', async () => {
-            const convex = new ConvexAPI(CONVEX_URL)
-            const importAccount = ConvexAccount.importFromString(PRIVATE_TEST_KEY_TEXT, PRIVATE_TEST_KEY_PASSWORD)
-            const account = await convex.createAccount(importAccount)
+            const convex = API.create(CONVEX_URL)
+            const importKeyPair = KeyPair.importFromString(PRIVATE_TEST_KEY_TEXT, PRIVATE_TEST_KEY_PASSWORD)
+            const account = await convex.createAccount(importKeyPair)
             const badLine = "(concat [*address*] 'test')"
             expect(
                 convex.query(badLine, account)
@@ -59,16 +60,16 @@ describe('ConvexAPI Class', () => {
         })
 
         it('should throw a ConvexAPIError on a new account balance', async () => {
-            const convex = new ConvexAPI(CONVEX_URL)
+            const convex = API.create(CONVEX_URL)
             const address = Number.MAX_SAFE_INTEGER
             const result = await convex.query(`(balance #${address})`, BigInt(9))
             assert.isNull(result['value'])
         })
 
         it('should return a valid checksum address from a query', async () => {
-            const convex = new ConvexAPI(CONVEX_URL)
-            const importAccount = ConvexAccount.importFromString(PRIVATE_TEST_KEY_TEXT, PRIVATE_TEST_KEY_PASSWORD)
-            const account = await convex.createAccount(importAccount)
+            const convex = API.create(CONVEX_URL)
+            const importKeyPair = KeyPair.importFromString(PRIVATE_TEST_KEY_TEXT, PRIVATE_TEST_KEY_PASSWORD)
+            const account = await convex.createAccount(importKeyPair)
             const address = account.address
             const result = await convex.query(`(address #${address})`, account)
             assert.equal(result['value'], account.address)
@@ -77,15 +78,17 @@ describe('ConvexAPI Class', () => {
 
     describe('getBalance', async () => {
         it('should get a 0 balance on a new account', async () => {
-            const convex = new ConvexAPI(CONVEX_URL)
-            const account = await convex.createAccount()
+            const convex = API.create(CONVEX_URL)
+            const keyPair = KeyPair.create()
+            const account = await convex.createAccount(keyPair)
             const value = await convex.getBalance(account)
             assert.equal(value, 0)
         })
 
         it('should get a small balance on a new account', async () => {
-            const convex = new ConvexAPI(CONVEX_URL)
-            const account = await convex.createAccount()
+            const convex = API.create(CONVEX_URL)
+            const keyPair = KeyPair.create()
+            const account = await convex.createAccount(keyPair)
             const amount = 100
             const result = await convex.requestFunds(amount, account)
             assert.equal(result, amount)
@@ -94,8 +97,9 @@ describe('ConvexAPI Class', () => {
         })
 
         it('should get a small balance on a new account using the address only field', async () => {
-            const convex = new ConvexAPI(CONVEX_URL)
-            const account = await convex.createAccount()
+            const convex = API.create(CONVEX_URL)
+            const keyPair = KeyPair.create()
+            const account = await convex.createAccount(keyPair)
             const amount = 100
             const result = await convex.requestFunds(amount, account)
             assert.equal(result, amount)
@@ -108,9 +112,9 @@ describe('ConvexAPI Class', () => {
         let convex
         let account
         before( async () => {
-            convex = new ConvexAPI(CONVEX_URL)
-            const importAccount = ConvexAccount.importFromString(PRIVATE_TEST_KEY_TEXT, PRIVATE_TEST_KEY_PASSWORD)
-            account = await convex.createAccount(importAccount)
+            convex = API.create(CONVEX_URL)
+            const importKeyPair = KeyPair.importFromString(PRIVATE_TEST_KEY_TEXT, PRIVATE_TEST_KEY_PASSWORD)
+            account = await convex.createAccount(importKeyPair)
             await convex.topupAccount(account)
         })
 
@@ -143,9 +147,9 @@ describe('ConvexAPI Class', () => {
         let functionName
         let functionAddress
         before( async () => {
-            convex = new ConvexAPI(CONVEX_URL)
-            const importAccount = ConvexAccount.importFromString(PRIVATE_TEST_KEY_TEXT, PRIVATE_TEST_KEY_PASSWORD)
-            account = await convex.createAccount(importAccount)
+            convex = API.create(CONVEX_URL)
+            const importKeyPair = KeyPair.importFromString(PRIVATE_TEST_KEY_TEXT, PRIVATE_TEST_KEY_PASSWORD)
+            account = await convex.createAccount(importKeyPair)
             await convex.topupAccount(account)
             functionName = 'test-storage'
             const testFunctionDeploy = `
@@ -180,53 +184,53 @@ describe('ConvexAPI Class', () => {
 
     describe('accountNames not resolved', async () => {
         let convex
-        let importAccount
+        let importKeyPair
         let accountName
         before( async () => {
-            convex = new ConvexAPI(CONVEX_URL)
-            importAccount = ConvexAccount.importFromString(PRIVATE_TEST_KEY_TEXT, PRIVATE_TEST_KEY_PASSWORD)
+            convex = API.create(CONVEX_URL)
+            importKeyPair = KeyPair.importFromString(PRIVATE_TEST_KEY_TEXT, PRIVATE_TEST_KEY_PASSWORD)
             accountName = 'test.convex-api.' + randomBytes(4).toString('hex')
         })
         it('should fail to reslove a new named account',  async() => {
-            const result = await convex.resolveAccountName(accountName, importAccount)
+            const result = await convex.resolveAccountName(accountName, importKeyPair)
             assert(!result)
         })
         it('should fail to load a new named account',  async() => {
-            const result = await convex.loadAccount(accountName, importAccount)
+            const result = await convex.loadAccount(accountName, importKeyPair)
             assert(!result)
         })
     })
 
     describe('accountNames register and load', async () => {
         let convex
-        let importAccount
+        let importKeyPair
         let accountName
         let newAccount
         before( async () => {
-            convex = new ConvexAPI(CONVEX_URL)
-            importAccount = ConvexAccount.importFromString(PRIVATE_TEST_KEY_TEXT, PRIVATE_TEST_KEY_PASSWORD)
+            convex = API.create(CONVEX_URL)
+            importKeyPair = KeyPair.importFromString(PRIVATE_TEST_KEY_TEXT, PRIVATE_TEST_KEY_PASSWORD)
             accountName = 'test.convex-api.' + randomBytes(4).toString('hex')
         })
         it('should setup and setup/create a new named account',  async() => {
-            newAccount = await convex.setupAccount(accountName, importAccount)
+            newAccount = await convex.setupAccount(accountName, importKeyPair)
             assert(newAccount)
             assert(newAccount.address)
             assert(newAccount.name)
         })
         it('should resolve the address of the new named account',  async() => {
-            const result = await convex.resolveAccountName(accountName, importAccount)
+            const result = await convex.resolveAccountName(accountName, importKeyPair)
             assert(result)
             assert.equal(result, newAccount.address)
         })
         it('should load the new named account',  async() => {
-            const result = await convex.loadAccount(accountName, importAccount)
+            const result = await convex.loadAccount(accountName, importKeyPair)
             assert(result)
             assert.equal(result.address, newAccount.address)
             assert.equal(result.name, newAccount.name)
         })
         it('should resolve the address of the new named account after a clear cache',  async() => {
             convex.registry.clearCache()
-            const result = await convex.resolveAccountName(accountName, importAccount)
+            const result = await convex.resolveAccountName(accountName, importKeyPair)
             assert(result)
             assert.equal(result, newAccount.address)
         })
@@ -235,7 +239,7 @@ describe('ConvexAPI Class', () => {
     describe('registry', async () => {
         let convex
         before( async () => {
-            convex = new ConvexAPI(CONVEX_URL)
+            convex = API.create(CONVEX_URL)
         })
         it('should resolve to a third party library or actor', async() => {
             const address = await convex.resolveName('convex.nft-tokens')
@@ -247,9 +251,9 @@ describe('ConvexAPI Class', () => {
         let convex
         let account
         before( async () => {
-            convex = new ConvexAPI(CONVEX_URL)
-            const importAccount = ConvexAccount.importFromString(PRIVATE_TEST_KEY_TEXT, PRIVATE_TEST_KEY_PASSWORD)
-            account = await convex.createAccount(importAccount)
+            convex = API.create(CONVEX_URL)
+            const importKeyPair = KeyPair.importFromString(PRIVATE_TEST_KEY_TEXT, PRIVATE_TEST_KEY_PASSWORD)
+            account = await convex.createAccount(importKeyPair)
             await convex.requestFunds(TOPUP_AMOUNT, account)
         })
         it('should reqeust information about an account object', async() => {
@@ -269,14 +273,15 @@ describe('ConvexAPI Class', () => {
         let convex
         let accountFrom
         before( async () => {
-            convex = new ConvexAPI(CONVEX_URL)
-            const importAccount = ConvexAccount.importFromString(PRIVATE_TEST_KEY_TEXT, PRIVATE_TEST_KEY_PASSWORD)
-            accountFrom = await convex.createAccount(importAccount)
+            convex = API.create(CONVEX_URL)
+            const importKeyPair = KeyPair.importFromString(PRIVATE_TEST_KEY_TEXT, PRIVATE_TEST_KEY_PASSWORD)
+            accountFrom = await convex.createAccount(importKeyPair)
             await convex.topupAccount(accountFrom)
         })
 
         it('should transfer a set amount of funds from the test account to a new account', async () => {
-            const accountTo = await convex.createAccount()
+            const keyPair = KeyPair.create()
+            const accountTo = await convex.createAccount(keyPair)
             const amount = BigInt(1000)
             const transferAmount = await convex.transfer(accountTo, amount, accountFrom)
             assert.equal(transferAmount, amount)
@@ -290,9 +295,9 @@ describe('ConvexAPI Class', () => {
         let convex
         let account
         before( async () => {
-            convex = new ConvexAPI(CONVEX_URL)
-            const importAccount = ConvexAccount.importFromString(PRIVATE_TEST_KEY_TEXT, PRIVATE_TEST_KEY_PASSWORD)
-            account = await convex.createAccount(importAccount)
+            convex = API.create(CONVEX_URL)
+            const importKeyPair = KeyPair.importFromString(PRIVATE_TEST_KEY_TEXT, PRIVATE_TEST_KEY_PASSWORD)
+            account = await convex.createAccount(importKeyPair)
             await convex.topupAccount(account)
         })
         it('should run multiple transactions before waiting for the result on the convex network', async () => {
@@ -314,17 +319,17 @@ describe('ConvexAPI Class', () => {
 
     describe('multi threaded createAccount test', async () => {
         let convex
-        let importAccount
+        let importKeyPair
         before( async () => {
-            convex = new ConvexAPI(CONVEX_URL)
-            importAccount = ConvexAccount.importFromString(PRIVATE_TEST_KEY_TEXT, PRIVATE_TEST_KEY_PASSWORD)
+            convex = API.create(CONVEX_URL)
+            importKeyPair = KeyPair.importFromString(PRIVATE_TEST_KEY_TEXT, PRIVATE_TEST_KEY_PASSWORD)
         })
         it('should run multiple createAccount test', async () => {
             const requestCount = 20
             let results = Array(requestCount)
             // first send transactions all at once
             for ( let counter = 0; counter < requestCount; counter ++ ) {
-                results[counter] = convex.createAccount(importAccount)
+                results[counter] = convex.createAccount(importKeyPair)
             }
             // now wait for the results
             for ( let counter = 0; counter < requestCount; counter ++ ) {